Move sources to named subdirectory
Change CMakeLists to detect if its included or used standalone
Change tests and sources to use proper source paths

+++ src/shl/file_stream.cpp
@@ -4,7 +4,7 @@
 
 #include <assert.h>
 #include <stdarg.h>
-#include "file_stream.hpp"
+#include "shl/file_stream.hpp"
 
 file_stream::operator FILE*() const
 {
@@ -97,21 +97,21 @@ bool is_ok(file_stream *stream)
         && (ferror(stream->handle) == 0);
 }
 
-size_t calculate_file_size(file_stream *stream)
+u64 calculate_file_size(file_stream *stream)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
 
-    size_t curpos = tell(stream);
+    u64 curpos = tell(stream);
     seek(stream, 0, SEEK_END);
-    size_t sz = tell(stream);
+    u64 sz = tell(stream);
     seek(stream, curpos);
 
     stream->size = sz;
     return sz;
 }
 
-size_t block_count(const file_stream *stream)
+u64 block_count(const file_stream *stream)
 {
     assert(stream != nullptr);
     return stream->size / stream->block_size;
@@ -129,30 +129,30 @@ int seek_block(file_stream *stream, long nth_block, int whence)
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
 
-    size_t bs = stream->block_size;
-    size_t block_pos = nth_block * bs;
+    u64 bs = stream->block_size;
+    u64 block_pos = nth_block * bs;
 
     if (whence != SEEK_CUR)
         return fseeko(stream->handle, block_pos, whence);
 
     // SEEK_CUR
-    size_t cur = (ftello(stream->handle) / bs) * bs;
+    u64 cur = (ftello(stream->handle) / bs) * bs;
     return fseeko(stream->handle, cur + block_pos, SEEK_SET);
 }
 
-int seek_next_alignment(file_stream *stream, size_t alignment)
+int seek_next_alignment(file_stream *stream, u64 alignment)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
     assert(alignment > 0);
 
-    size_t npos = tell(stream);
+    u64 npos = tell(stream);
     npos = (npos + alignment - 1) / alignment * alignment;
 
     return fseeko(stream->handle, npos, SEEK_SET);
 }
 
-size_t tell(file_stream *stream)
+u64 tell(file_stream *stream)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -177,7 +177,7 @@ bool rewind(file_stream *stream)
     return fseeko(stream->handle, 0L, SEEK_SET) == 0;
 }
 
-size_t read(file_stream *stream, void *out, size_t size)
+u64 read(file_stream *stream, void *out, u64 size)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -186,7 +186,7 @@ size_t read(file_stream *stream, void *out, size_t size)
     return fread(out, 1, size, stream->handle);
 }
 
-size_t read(file_stream *stream, void *out, size_t size, size_t nmemb)
+u64 read(file_stream *stream, void *out, u64 size, u64 nmemb)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -195,7 +195,7 @@ size_t read(file_stream *stream, void *out, size_t size, size_t nmemb)
     return fread(out, size, nmemb, stream->handle);
 }
 
-size_t read_at(file_stream *stream, void *out, size_t offset, size_t size)
+u64 read_at(file_stream *stream, void *out, u64 offset, u64 size)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -207,7 +207,7 @@ size_t read_at(file_stream *stream, void *out, size_t offset, size_t size)
     return fread(out, 1, size, stream->handle);
 }
 
-size_t read_at(file_stream *stream, void *out, size_t offset, size_t size, size_t nmemb)
+u64 read_at(file_stream *stream, void *out, u64 offset, u64 size, u64 nmemb)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -219,7 +219,7 @@ size_t read_at(file_stream *stream, void *out, size_t offset, size_t size, size_
     return fread(out, size, nmemb, stream->handle);
 }
 
-size_t read_block(file_stream *stream, void *out)
+u64 read_block(file_stream *stream, void *out)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -228,7 +228,7 @@ size_t read_block(file_stream *stream, void *out)
     return fread(out, 1, stream->block_size, stream->handle);
 }
 
-size_t read_block(file_stream *stream, void *out, size_t nth_block)
+u64 read_block(file_stream *stream, void *out, u64 nth_block)
 {
     assert(out != nullptr);
     
@@ -236,7 +236,7 @@ size_t read_block(file_stream *stream, void *out, size_t nth_block)
     return fread(out, 1, stream->block_size, stream->handle);
 }
 
-size_t read_blocks(file_stream *stream, void *out, size_t block_count)
+u64 read_blocks(file_stream *stream, void *out, u64 block_count)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -245,7 +245,7 @@ size_t read_blocks(file_stream *stream, void *out, size_t block_count)
     return fread(out, stream->block_size, block_count, stream->handle);
 }
 
-size_t read_blocks(file_stream *stream, void *out, size_t nth_block, size_t block_count)
+u64 read_blocks(file_stream *stream, void *out, u64 nth_block, u64 block_count)
 {
     assert(out != nullptr);
 
@@ -253,27 +253,27 @@ size_t read_blocks(file_stream *stream, void *out, size_t nth_block, size_t bloc
     return fread(out, stream->block_size, block_count, stream->handle);
 }
 
-size_t read_entire_file(file_stream *stream, void *out, size_t max_size)
+u64 read_entire_file(file_stream *stream, void *out, u64 max_size)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
     assert(out != nullptr);
     
-    size_t old_pos = ftello(stream->handle);
+    u64 old_pos = ftello(stream->handle);
     fseeko(stream->handle, 0, SEEK_SET);
 
-    size_t sz = stream->size;
+    u64 sz = stream->size;
 
     if (sz > max_size)
         sz = max_size;
 
-    size_t ret = fread(out, 1, sz, stream->handle);
+    u64 ret = fread(out, 1, sz, stream->handle);
     fseeko(stream->handle, old_pos, SEEK_SET);
 
     return ret;
 }
 
-size_t write(file_stream *stream, const void *in, size_t size)
+u64 write(file_stream *stream, const void *in, u64 size)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -282,7 +282,7 @@ size_t write(file_stream *stream, const void *in, size_t size)
     return fwrite(in, 1, size, stream->handle);
 }
 
-size_t write(file_stream *stream, const void *in, size_t size, size_t nmemb)
+u64 write(file_stream *stream, const void *in, u64 size, u64 nmemb)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -291,7 +291,7 @@ size_t write(file_stream *stream, const void *in, size_t size, size_t nmemb)
     return fwrite(in, size, nmemb, stream->handle);
 }
 
-size_t write(file_stream *stream, const char *str)
+u64 write(file_stream *stream, const char *str)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -300,7 +300,7 @@ size_t write(file_stream *stream, const char *str)
     return fputs(str, stream->handle);
 }
 
-size_t write_at(file_stream *stream, const void *in, size_t offset, size_t size)
+u64 write_at(file_stream *stream, const void *in, u64 offset, u64 size)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -312,7 +312,7 @@ size_t write_at(file_stream *stream, const void *in, size_t offset, size_t size)
     return fwrite(in, 1, size, stream->handle);
 }
 
-size_t write_at(file_stream *stream, const void *in, size_t offset, size_t size, size_t nmemb)
+u64 write_at(file_stream *stream, const void *in, u64 offset, u64 size, u64 nmemb)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -324,7 +324,7 @@ size_t write_at(file_stream *stream, const void *in, size_t offset, size_t size,
     return fwrite(in, size, nmemb, stream->handle);
 }
 
-size_t write_at(file_stream *stream, const char *str, size_t offset)
+u64 write_at(file_stream *stream, const char *str, u64 offset)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -336,7 +336,7 @@ size_t write_at(file_stream *stream, const char *str, size_t offset)
     return fputs(str, stream->handle);
 }
 
-size_t write_block(file_stream *stream, const void *in)
+u64 write_block(file_stream *stream, const void *in)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -345,7 +345,7 @@ size_t write_block(file_stream *stream, const void *in)
     return fwrite(in, 1, stream->block_size, stream->handle);
 }
 
-size_t write_block(file_stream *stream, const void *in, size_t nth_block)
+u64 write_block(file_stream *stream, const void *in, u64 nth_block)
 {
     assert(in != nullptr);
 
@@ -353,7 +353,7 @@ size_t write_block(file_stream *stream, const void *in, size_t nth_block)
     return fwrite(in, 1, stream->block_size, stream->handle);
 }
 
-size_t write_blocks(file_stream *stream, const void *in, size_t block_count)
+u64 write_blocks(file_stream *stream, const void *in, u64 block_count)
 {
     assert(stream != nullptr);
     assert(stream->handle != nullptr);
@@ -362,7 +362,7 @@ size_t write_blocks(file_stream *stream, const void *in, size_t block_count)
     return fwrite(in, stream->block_size, block_count, stream->handle);
 }
 
-size_t write_blocks(file_stream *stream, const void *in, size_t nth_block, size_t block_count)
+u64 write_blocks(file_stream *stream, const void *in, u64 nth_block, u64 block_count)
 {
     assert(in != nullptr);
